SecureVoting/requirements.txt
Django>=4.2.0
opencv-python>=4.8.0
ImageHash>=4.3.1
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0